### 2022-11-16 #4941 @editorialbot
[PRE REVIEW]: ALNS: a Python implementation of the adaptive large neighbourhood search metaheuristic
**Submitting author:** @N-Wouda (<a href="http://orcid.org/0000-0003-2463-0309">Niels Wouda</a>)
**Repository:** https://github.com/N-Wouda/ALNS
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** v5.0.0
