### 2022-11-17 a2a76d2
Add method argument to SimulatedAnnealing.autofit
