+++ alns/accept/tests/test_simulated_annealing.py
@@ -60,12 +60,21 @@ def test_raises_start_smaller_than_end():
     SimulatedAnnealing(1, 1, 1)  # should not raise for equality
 
 
+def test_raises_unknown_method():
+    with assert_raises(ValueError):
+        # Accepts only 'exponential' or 'linear'
+        SimulatedAnnealing(10, 5, 1, "unknown method")
+
+
 def test_does_not_raise():
     """
     These sets of parameters should work correctly.
     """
     SimulatedAnnealing(10, 5, 1, "exponential")
+    SimulatedAnnealing(10, 5, 1, "EXPONENTIAL")
+
     SimulatedAnnealing(10, 5, 2, "linear")
+    SimulatedAnnealing(10, 5, 2, "LINEAR")
 
 
 @mark.parametrize("step", range(10))
@@ -171,7 +180,8 @@ def test_accepts_generator_and_random_state():
         (-1, 0.5, 10),  # negative worse
         (0, -1, 10),  # negative prob
         (1.5, 0.5, 10),  # worse outside unit interval
-        (1, 0.9, -10),
+        (1, 0.9, -10),  # negative iterations
+        (1, 0.9, 0),  # zero iterations
     ],
 )  # negative number of iterations
 def test_autofit_raises_for_invalid_inputs(
@@ -206,3 +216,13 @@ def test_autofit_on_several_examples(
     assert_almost_equal(sa.end_temperature, sa_end)
     assert_almost_equal(sa.step, sa_step)
     assert_equal(sa.method, "exponential")
+
+
+def test_linear_autofit():
+    sa = SimulatedAnnealing.autofit(100, 0.05, 0.5, 100, "linear")
+    sa_start = -0.05 * 100 / np.log(0.5)
+
+    assert_almost_equal(sa.start_temperature, sa_start)
+    assert_almost_equal(sa.end_temperature, 1)
+    assert_almost_equal(sa.step, (sa_start - 1) / 100)
+    assert_equal(sa.method, "linear")
