### 2022-11-18 cccd5ab
Streamline RRT/SA autofit a bit more
