### 2022-11-18 f3e717e
Test unknown method raises ValueError
