### 2022-11-18 795f700
Add method argument to SimulatedAnnealing.autofit (#118)
* Add method argument to SimulatedAnnealing.autofit

* Streamline RRT/SA autofit a bit more
