+++ alns/accept/SimulatedAnnealing.py
@@ -102,7 +102,12 @@ class SimulatedAnnealing(AcceptanceCriterion):
 
     @classmethod
     def autofit(
-        cls, init_obj: float, worse: float, accept_prob: float, num_iters: int
+        cls,
+        init_obj: float,
+        worse: float,
+        accept_prob: float,
+        num_iters: int,
+        method: str = "exponential",
     ) -> "SimulatedAnnealing":
         """
         Returns an SA object with initial temperature such that there is a
@@ -118,20 +123,22 @@ class SimulatedAnnealing(AcceptanceCriterion):
         init_obj
             The initial solution objective.
         worse
-            Percentage (between 0 and 1) the candidate solution may be worse
-            than initial solution for it to be accepted with probability
+            Percentage (in (0, 1), exclusive) the candidate solution may be
+            worse than initial solution for it to be accepted with probability
             ``accept_prob``.
         accept_prob
-            Initial acceptance probability for a solution at most ``worse``
-            worse than the initial solution.
+            Initial acceptance probability (in [0, 1]) for a solution at most
+            ``worse`` worse than the initial solution.
         num_iters
             Number of iterations the ALNS algorithm will run.
+        method
+            The updating method, one of {'linear', 'exponential'}. Default
+            'exponential'.
 
         Raises
         ------
         ValueError
-            When ``worse`` not in [0, 1] or when ``accept_prob`` is not in
-            (0, 1).
+            When the parameters do not meet requirements.
 
         Returns
         -------
@@ -154,12 +161,22 @@ class SimulatedAnnealing(AcceptanceCriterion):
         if not (0 < accept_prob < 1):
             raise ValueError("accept_prob outside (0, 1) not understood.")
 
-        if num_iters < 0:
-            raise ValueError("Negative number of iterations not understood.")
+        if num_iters <= 0:
+            raise ValueError("Non-positive num_iters not understood.")
+
+        if method not in ["linear", "exponential"]:
+            raise ValueError("Method must be one of ['linear', 'exponential']")
 
         start_temp = -worse * init_obj / np.log(accept_prob)
-        step = (1 / start_temp) ** (1 / num_iters)
 
-        logger.info(f"Autofit start_temp {start_temp:.2f}, step {step:.2f}.")
+        if method == "linear":
+            step = (start_temp - 1) / num_iters
+        else:
+            step = (1 / start_temp) ** (1 / num_iters)
 
-        return cls(start_temp, 1, step, method="exponential")
+        logger.info(
+            f"Autofit {method} SA: start_temp {start_temp:.2f}, "
+            f"step {step:.2f}."
+        )
+
+        return cls(start_temp, 1, step, method=method)
