### 2022-11-19 14e9ad4
Always run CI matrix
