+++ .github/workflows/alns.yaml
@@ -20,11 +20,14 @@ jobs:
         uses: actions/setup-python@v4
         with:
           python-version: ${{ matrix.python-version }}
+      - name: Get current time
+        uses: josStorer/get-current-time@v2.0.2
+        id: current-time
       - uses: actions/cache@v3
         id: cache
         with:
           path: ${{ env.pythonLocation }}
-          key: ${{ runner.os }}-python-${{ env.pythonLocation }}-${{ hashFiles('pyproject.toml') }}-$(date +'%Y-%m')
+          key: ${{ runner.os }}-python-${{ env.pythonLocation }}-${{ hashFiles('pyproject.toml') }}-${{ steps.current-time.outputs.year }}-${{ steps.current-time.outputs.month }}
       - name: Install dependencies
         if: steps.cache.outputs.cache-hit != 'true'
         run: |
