+++ .github/workflows/alns.yaml
@@ -23,18 +23,27 @@ jobs:
       - name: Get current time
         uses: josStorer/get-current-time@v2.0.2
         id: current-time
-      - uses: actions/cache@v3
-        id: cache
+      - name: Cache Python dependencies
+        uses: actions/cache@v3
+        id: cache-python
         with:
           path: ${{ env.pythonLocation }}
           key: ${{ runner.os }}-python-${{ env.pythonLocation }}-${{ hashFiles('pyproject.toml') }}-${{ steps.current-time.outputs.year }}-${{ steps.current-time.outputs.month }}
-      - name: Install dependencies
-        if: steps.cache.outputs.cache-hit != 'true'
+      - name: Install Python dependencies
+        if: steps.cache-python.outputs.cache-hit != 'true'
         run: |
           python -m pip install --upgrade pip
           pip install poetry
           poetry install
-          poetry run pre-commit install
+      - name: Cache pre-commit
+        uses: actions/cache@v3
+        id: cache-pre-commit
+        with:
+          path: ~/.cache/pre-commit/
+          key: pre-commit-4|${{ env.pythonLocation }}|${{ hashFiles('.pre-commit-config.yaml') }}
+      - name: Install pre-commit
+        if: steps.cache-pre-commit.outputs.cache-hit != 'true'
+        run: poetry run pre-commit install
       - name: Run pytest
         run: poetry run pytest
       - name: Run pre-commit
