### 2022-11-19 137baf9
Simplify pre-commit hash key
