+++ .github/workflows/alns.yaml
@@ -27,7 +27,7 @@ jobs:
         uses: actions/cache@v3
         id: cache-python
         with:
-          path: ${{ env.pythonLocation }}
+          path: ~/.cache/poetry
           key: ${{ runner.os }}-python-${{ env.pythonLocation }}-${{ hashFiles('pyproject.toml') }}-${{ steps.current-time.outputs.year }}-${{ steps.current-time.outputs.month }}
       - name: Install Python dependencies
         if: steps.cache-python.outputs.cache-hit != 'true'
