+++ .github/workflows/alns.yaml
@@ -20,6 +20,10 @@ jobs:
         uses: actions/setup-python@v4
         with:
           python-version: ${{ matrix.python-version }}
+      - name: Update pip and poetry
+        run: |
+          python -m pip install --upgrade pip
+          pip install poetry
       - name: Get current time
         uses: josStorer/get-current-time@v2.0.2
         id: current-time
@@ -27,14 +31,11 @@ jobs:
         uses: actions/cache@v3
         id: cache-python
         with:
-          path: ~/.cache/poetry
+          path: ~/.cache/pypoetry
           key: ${{ runner.os }}-python-${{ env.pythonLocation }}-${{ hashFiles('pyproject.toml') }}-${{ steps.current-time.outputs.year }}-${{ steps.current-time.outputs.month }}
       - name: Install Python dependencies
         if: steps.cache-python.outputs.cache-hit != 'true'
-        run: |
-          python -m pip install --upgrade pip
-          pip install poetry
-          poetry install
+        run: poetry install
       - name: Cache pre-commit
         uses: actions/cache@v3
         id: cache-pre-commit
@@ -59,7 +60,7 @@ jobs:
         uses: actions/setup-python@v4
         with:
           python-version: 3.9
-      - name: Install dependencies
+      - name: Update pip and poetry
         run: |
           python -m pip install --upgrade pip
           pip install poetry
