### 2022-11-19 dcff19a
Tinker with caching in CI (#119)
* Bump actions/setup-python

This fixes a deprecation warning in the CI runner.

* Separate pre-commit cache

* Cache poetry

* Fix poetry cache location, move update into separate step
