+++ .github/workflows/alns.yaml
@@ -5,10 +5,8 @@ on:
     branches: [ master ]
   pull_request:
     branches: [ master ]
-    paths:
-      - 'alns/**.py'
   release:
-    types: [created]
+    types: [ created ]
 
 jobs:
   build:
@@ -19,15 +17,34 @@ jobs:
     steps:
       - uses: actions/checkout@v3
       - name: Set up Python
-        uses: actions/setup-python@v3
+        uses: actions/setup-python@v4
         with:
           python-version: ${{ matrix.python-version }}
-      - name: Install dependencies
+      - name: Update pip and poetry
         run: |
           python -m pip install --upgrade pip
           pip install poetry
-          poetry install
-          poetry run pre-commit install
+      - name: Get current time
+        uses: josStorer/get-current-time@v2.0.2
+        id: current-time
+      - name: Cache Python dependencies
+        uses: actions/cache@v3
+        id: cache-python
+        with:
+          path: ~/.cache/pypoetry
+          key: ${{ runner.os }}-python-${{ env.pythonLocation }}-${{ hashFiles('pyproject.toml') }}-${{ steps.current-time.outputs.year }}-${{ steps.current-time.outputs.month }}
+      - name: Install Python dependencies
+        if: steps.cache-python.outputs.cache-hit != 'true'
+        run: poetry install
+      - name: Cache pre-commit
+        uses: actions/cache@v3
+        id: cache-pre-commit
+        with:
+          path: ~/.cache/pre-commit/
+          key: pre-commit-${{ env.pythonLocation }}-${{ hashFiles('.pre-commit-config.yaml') }}
+      - name: Install pre-commit
+        if: steps.cache-pre-commit.outputs.cache-hit != 'true'
+        run: poetry run pre-commit install
       - name: Run pytest
         run: poetry run pytest
       - name: Run pre-commit
@@ -40,10 +57,10 @@ jobs:
     steps:
       - uses: actions/checkout@v3
       - name: Set up Python
-        uses: actions/setup-python@v3
+        uses: actions/setup-python@v4
         with:
           python-version: 3.9
-      - name: Install dependencies
+      - name: Update pip and poetry
         run: |
           python -m pip install --upgrade pip
           pip install poetry
