### 2022-11-19 fc26da9
Add note with reference to original to notebooks in docs, update doc structure, switch SA to full math docs
