### 2022-11-19 bec0aa3
Add math to WA and RRT docs
