### 2022-11-19 033c84e
Improve docs (#120)
* Add note with reference to original to notebooks in docs, update doc structure

* Add math to SA docs

* Add math to WA and RRT docs
