### 2022-11-19 31d4bee
Move config from pre-commit to pyproject.toml and setup.cfg (#121)
