+++ .pre-commit-config.yaml
@@ -10,22 +10,18 @@ repos:
     rev: 22.3.0
     hooks:
     -   id: black
-        args: [--line-length=79]
 
 -   repo: https://github.com/pycqa/flake8
     rev: 4.0.1
     hooks:
     -   id: flake8
-        args: [--per-file-ignores=__init__.py:F401, --extend-ignore="E203"]
 
 -   repo: https://github.com/pycqa/isort
     rev: 5.10.1
     hooks:
     -   id: isort
-        args: [--case-sensitive, --profile=black, --line-length=79]
 
 -   repo: https://github.com/pre-commit/mirrors-mypy
     rev: v0.982
     hooks:
     -   id: mypy
-        args: [--ignore-missing-imports]
