### 2022-11-19 e4bac14
Add DOI and journal info for Wouda et al. 2022
