+++ paper/paper.bib
@@ -57,12 +57,15 @@
     year = {2022},
 }
 
-@misc{Wouda_et_al:2022,
-    author = {Wouda, Niels A. and Ayse Aslan and Iris F. A. Vis},
-    title = {An adaptive large neighbourhood search metaheuristic for hourly learning activity planning in personalised learning},
-    year = {2022},
-    howpublished = {[Submitted for publication]},
-    url = {https://github.com/N-Wouda/PL-Hourly-Heuristic},
+@article{Wouda_et_al:2022,
+    title={An adaptive large neighbourhood search metaheuristic for hourly learning activity planning in personalised learning},
+    ISSN={0305-0548},
+    DOI={10.1016/j.cor.2022.106089},
+    abstractNote={Personalised learning offers an alternative method to one-size-fits-all education in schools, and has seen increasing adoption over the past several years. Personalised learning’s focus on learner-driven education requires novel scheduling methods. In this paper we introduce the hourly, learner-driven activity planning problem of personalised learning, and formulate scheduling methods to solve it. We present an integer linear programming model of the problem, but this model does not generate schedules sufficiently quickly for use in practice. To overcome this, we propose an adaptive large neighbourhood search metaheuristic to solve the problem instead. The metaheuristic’s performance is compared against optimal solutions in a large numerical study of 14,400 instances. These instances are representative of secondary education in the Netherlands, and were developed from expert opinions. Solutions on average deviate only 1.6% from optimal results. Further, our experiments numerically demonstrate the mitigating effects changes to the structure and staffing of secondary education have on the challenges of satisfying learner instruction demands in personalised learning.},
+    journal={Computers & Operations Research},
+    author={Wouda, Niels A. and Aslan, Ayse and Vis, Iris F. A.},
+    year={2022},
+    pages={106089},
 }
 
 @article{vidal:2022,
