### 2022-11-23 7f46368
Add mathematical descriptions to SA, RouletteWheel, and SegmentedRouletteWheel docs
