### 2022-11-23 0f167a7
Add statement of need to index page, stubs for installation and contributing
