+++ docs/source/setup/installation.rst
@@ -1,0 +1,13 @@
+Installation instructions
+=========================
+
+The most straightforward way to use the ``alns`` package in your project is to install via *pip*, like so:
+
+.. code-block:: shell
+
+   pip install alns
+
+
+TODO from source
+
+TODO running the examples
