+++ docs/source/setup/getting_help.rst
@@ -1,0 +1,4 @@
+Getting help
+============
+
+TODO
