### 2022-11-23 1623862
Move notebooks to examples/ folder, describe how to install and run everything
