### 2022-11-23 aa45166
Copy over examples in conf.py
