+++ docs/requirements.txt
@@ -4,4 +4,3 @@ tomli >= 2.0.1
 nbsphinx >= 0.8.9
 ipython >= 8.6.0
 numpydoc >= 1.5.0
-sphinx-collections >= 0.0.1
