### 2022-11-23 028aa9e
Fix typo's and vagueness
