### 2022-11-23 cf3412d
Improve documentation (#124)
* Add mathematical descriptions to SA, RouletteWheel, and SegmentedRouletteWheel docs

* Add statement of need to index page, stubs for installation and contributing

* Move notebooks to examples/ folder, describe how to install and run everything

* Copy over examples in conf.py

* Finish getting help page

* Write about how to contribute

* Update bug report template

* Fix typo's and vagueness
