+++ docs/source/conf.py
@@ -1,6 +1,9 @@
 import datetime
+import glob
 import os
+import shutil
 import sys
+from pathlib import Path
 
 import tomli
 
@@ -18,6 +21,12 @@ with open("../../pyproject.toml", "rb") as fh:
     pyproj = tomli.load(fh)
     release = version = pyproj["tool"]["poetry"]["version"]
 
+for file in glob.iglob("../../examples/*.ipynb"):
+    path = Path(file)
+
+    print(f"Copy {path.name} into docs/source/examples/")
+    shutil.copy2(path, f"examples/{path.name}")
+
 # -- Autodoc
 
 autoclass_content = "class"
@@ -34,26 +43,6 @@ numpydoc_class_members_toctree = False
 # -- nbsphinx
 
 nbsphinx_execute = "never"
-
-nbsphinx_prolog = r"""
-{% set docname = 'docs/source/' + env.doc2path(env.docname, base=None) %}
-{% set url_prefix = 'https://github.com/N-Wouda/ALNS/blob/' %}
-{% set release = env.config.release %}
-
-.. raw:: html
-
-   <div class="admonition note">
-     <p class="admonition-title">
-        Note
-     </p>
-     <p>
-        This page was generated from
-        <a class="reference external"
-           href="{{ url_prefix|e }}v{{ release|e }}/{{ docname|e }}">
-           {{ docname|e }}</a>.
-     </p>
-   </div>
-"""
 
 # -- General configuration
 
