### 2022-11-24 bfa6f5c
Add code of conduct based on contributor covenant
And update my public email
