+++ pyproject.toml
@@ -2,7 +2,7 @@
 name = "alns"
 version = "5.0.1"
 description = "A flexible implementation of the adaptive large neighbourhood search (ALNS) algorithm."
-authors = ["Niels Wouda <n.wouda@apium.nl>"]
+authors = ["Niels Wouda <nielswouda@gmail.com>"]
 license = "MIT"
 readme = "README.md"
 repository = "https://github.com/N-Wouda/ALNS"
