+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "alns"
-version = "5.0.1"
+version = "5.0.2"
 description = "A flexible implementation of the adaptive large neighbourhood search (ALNS) algorithm."
 authors = ["Niels Wouda <nielswouda@gmail.com>"]
 license = "MIT"
