### 2022-11-24 #4941 @N-Wouda
@danielskatz @hugoledoux can we somehow update the version to v5.0.2? My co-author and I have added a few things over the past weeks that caused the text (and PyPI package) to diverge a bit from the version I originally put up for review.
