### 2022-11-25 #4941 @hugoledoux
yes you can, no problems.

I'm looking for reviewers, it should take a few days, and then the review starts. But part of the review is to improve the code and eventually release a new version, so updating now is not an issue at all 👍
