### 2022-11-25 8ce7209
Remove phrasing about returned solutions not being evaluated again
That's weird anyway, because the callback signature is that it does not return. The phrasing was a bit of a historical artifact that we should not have kept around so long.
