### 2022-11-25 3234398
Closes #126: adds a list of publications to the docs
The list's pretty empty for now, but should see a few more contributions soon enough!
