+++ docs/source/misc/publications.rst
@@ -1,0 +1,15 @@
+Publications using ``alns``
+===========================
+
+The following publications made use of the ``alns`` package in some way.
+
+.. note::
+
+    Want to get your own publication in this list?
+    Feel free to open an issue in the `GitHub repository <https://github.com/N-Wouda/ALNS/>`_ with the publication details, or directly open a pull request there.
+    Do make sure to include the publication's DOI!
+
+2022
+----
+
+Wouda, N.A., A. Aslan, and I.F.A. Vis (2022). An adaptive large neighbourhood search metaheuristic for hourly learning activity planning in personalised learning. *Computers & Operations Research*: 106089. https://doi.org/10.1016/j.cor.2022.106089.
