### 2022-12-15 a37e484
Simplify readme a bit, remove misc/publications (#128)
