### 2022-12-19 #4941 @skadio
Understand. There is a small window of oppt. this week as things are slowing down to get this review going. After that, the break comes follow-up with a heavy start of the year. The next window I would be looking at would be later in Q1.
