### 2022-12-25 #5028 @skadio
I played with this library, and first things first. I LOVE it! Kudos to @N-Wouda and the team!

The summary of my JOSS review, minimum action items needed for publication, and my detailed commentary is below.

- This paper and the library would an excellent contribution to the community and JOSS
-  In the big picture, it can help technical but non-optimization-expert users to jump start with very effective, practical approaches via ALNS

At a minimum, we need:

1. The notebooks and documentations are great! Someone with an OR background should not have any problem using the library. That said, we need a bit more on the example usage for outsiders. With a quick low-hanging effort, the library can target a larger audience --which I believe is where largest potential is. To inspire the authors, I took a stab at this below. More on this in details below.

2. See the JOSS required item above to please add "State of the Field" to cover how this package compares to other packages. If none such exists (which I doubt it, since stochastic local search is well-known), please make the contribution explicit. If there are other tools, please cover mainstream ones, if any and mention where each would benefit the most (e.g., genetic algorithms for multi-point solutions vs. alns)

3. See the JOSS required item above to please add Community Guidelines, if not already.
