### 2022-12-25 #5028 @skadio
**Detailed comments:**

My main concern is that (for a non-OR user), the "minimum" set of requirements is not clearly defined. What does the user need to bring to the table to benefit from this library.

The current README only provides a quick comment, and while the Documentations is quite nice/detailed, it does not provide a crisp answer to a newcomer.

Any optimization expert should be able to consume the library but with JOSS, there is a potential for a larger audience (who are technical experts, say a Kaggler who wants to use this library in a competition, just not familiar with optimization know-how/jargon).

What would help the most is a **high-level, problem independent boilerplate implementation with TODO statements** to guide the user to fill for their first ALNS model. Specific problem implementations in the documents can then become extremely useful for detailed applications.

Here is one example boiler plate. I am sharing it here, so that the authors can, first of all, correct me if my understanding of the library is incorrect.

I _believe_, at a MINIMUM, the library requires the user to define:

1. a STATE
2. at least one DESTROY and one REPAIR operator, and
3. an INITIAL STATE.

Is this correct? If so, here is my boiler plate:
