### 2022-12-25 #5028 @skadio
```
import numpy.random as rnd
from alns import ALNS, State
from alns.select import RouletteWheel
from alns.accept import HillClimbing
from alns.stop import MaxRuntime

class MyState(State):
    """
    Problem specific solution state
    """

    def __init__(self, my_parameter):
        self.my_parameter = my_parameter

    def objective(self) -> float:
        """
        This function calculates the objective of the state.
        Remember, ALNS works minimization problems!
        """

	# TODO implement the objective function
        return NotImplementedError()

def create_initial_state() -> MyState:
	# TODO implement the creation of an initial state and return it

	# Create a state and setup the initial solution
	state = MyState("My problem specific state")

	return NotImplementedError()

def my_state_destroy(state: MyState) -> MyState:
	# TODO: implement how destroy a given state and return the destroyed state

	# Remember to destroy a copy!
	destroyed = copy.deepcopy(state)

	return NotImplementedError()

def my_state_repair(state: MyState) -> MyState:

	# TODO: implement how repair a given state and return repaired state
	return NotImplementedError()

# Create ALNS
alns = ALNS(rnd.RandomState(seed=42))

# Add one or more destroy and repair operators to alns
alns.add_destroy_operator(my_state_destroy)
alns.add_repair_operator(my_state_repair)

# Configurate ALNS
select = RouletteWheel()	# See alns.select for others
accept = HillClimbing()		# See alns.accept for others
stop = MaxRuntime(60)		# See alns.stop for others

# Create the initial solution
initial_solution = create_initial_state()
print("Initial solution objective is ", initial_solution.objective()")

# Run ALNS
result = alns.iterate(initial_solution, select, accept, stop)

# Retrieve the final solution
solution = result.best_state
objective = solution.objective()
print("Best heuristic objective is:", objective)

```

You can correct this as needed, hope the idea is clear. I tried to add useful commentary, type hints etc. to guide the user. Please extend from there.

In the README:

- You shall have a LIST of what are the MINIMUM TODO items a user needs.
- Then, you can present a boiler plate, like the above.

In the PAPER:
- I would again recommend the minimum to-do list,
- and even the boiler plate there, with much less commentary, type hints etc. where the idea is to give a flavor for the library usage example. Anyone reading the paper would then be excited to try out their ideas (seeing how easy it would be to solve their problems at hand)
