### 2022-12-25 #5028 @skadio
Detailed Comments Continued:

Some minor code/low-level commentaries, mostly for my curiosity:

- The pydoc in State.py class reads "you are _encouraged_ to subclass it". My understanding is, doing is MANDATORY not encourage. If so, please re-word. Same everywhere to be explicit about what is mandatory.
- In the Destroy() operator is it mandatory to work with the deep copy object? It seems so to me, but if correct, please be explicit about it since this is critical, no? Again, this is where the boiler plate comes quite handy. I explicitly mentioned this in the code TODO above
- If I am following correctly, the library is built for minimization. Again, this is quite critical and needs to be explicitly stated. See my TODO comments about this and aim for something similar
- I saw some use of @dataclass annotation in some of the STATE definitions. Why is this needed/useful? and when? Documentation does not say much about its use. Please add some color
- In the source code /examples folder, if you create a /data folder, then you can move all data files there. This would clean-up the structure and keep the notebooks (the most interesting bit) much more visible to library users.
- On that note, here is nice to have comment, and completely not necessary, that you are also probably aware of: there is A LOT of great code used in the documentation. At the same time, there is an opportunity to have them built-in within the library. I am thinking of sklearn and other ML library that comes ready with data loaders, plotting classes, that can be imported and used directly. If these data elements were moved into the library, then you could drop all that code from the documentation and hugely simplify the ALNS presentation with nice imports (say draw_tsp()), and the purpose is clear from the context. This would also hide a lot non-ALNS related imports/configs from sight as well. Again, this is not critical at all. Just mentioning, and I am sure you already considered.
