### 2022-12-25 #5028 @skadio
To summarize: great library/paper for publication at JOSS. Only a few minimum items to revise, which should be easy for the team.

Thank you for the great work! @N-Wouda
