### 2022-12-26 N-Wouda/ALNS#129 @N-Wouda
"State of the Field" section in JOSS paper
From the review in https://github.com/openjournals/joss-reviews/issues/5028:

>State of the field: Do the authors describe how this software compares to other commonly-used packages?

And reviewer comment:

>See the JOSS required item above to please add "State of the Field" to cover how this package compares to other packages. If none such exists (which I doubt it, since stochastic local search is well-known), please make the contribution explicit. If there are other tools, please cover mainstream ones, if any and mention where each would benefit the most (e.g., genetic algorithms for multi-point solutions vs. alns)

There's some references already written down in the review that might be useful. I'm thinking ten-or-so references at most (incl. some open source Python/other optimization packages), and text similar in size to the statement of need.

@leonlan can you have a go at this?
