### 2022-12-26 N-Wouda/ALNS#130 @N-Wouda
Simple, cookie-cutter template to get started
From review in https://github.com/openjournals/joss-reviews/issues/5028:

>The notebooks and documentations are great! Someone with an OR background should not have any problem using the library. That said, we need a bit more on the example usage for outsiders. With a quick low-hanging effort, the library can target a larger audience --which I believe is where largest potential is. To inspire the authors, I took a stab at this below. More on this in details below.

We need to provide a simple example document to explain to new users what they need to bring to the table to get started using the package. There's some example boilerplate available in the review issue. We should adopt that and feature it prominently throughout the package.
