### 2022-12-27 cddf6b4
Add a "getting started" code template
