### 2022-12-27 c45b7be
Move data files into examples/data, update notebooks
