+++ docs/source/index.rst
@@ -2,12 +2,12 @@ ALNS
 ====
 
 ``alns`` is a general, well-documented and tested implementation of the adaptive large neighbourhood search (ALNS) metaheuristic in Python.
-The ALNS metaheuristic is an algorithm that can be used to solve difficult combinatorial optimisation problems.
+ALNS is an algorithm that can be used to solve difficult combinatorial optimisation problems.
 The algorithm begins with an initial solution.
 This solution should be feasible, but need not be very good.
 Then the algorithm iterates until a stopping criterion is met.
-In each iteration, it selects a destroy and repair operator, which transform the current solution into a candidate solution.
-This candidate solution is then evaluated by an acceptance criterion, and the operator selection mechanism is updated based on the evaluation outcome.
+In each iteration, a destroy and repair operator are selected, which transform the current solution into a candidate solution.
+This candidate solution is then evaluated by an acceptance criterion, and the operator selection scheme is updated based on the evaluation outcome.
 
 It is common in the operations research community to re-implement heuristics such as ALNS.
 Such implementations are relatively limited, are typically tied tightly to one particular problem domain, and often implement just a single acceptance criterion and operator selection scheme.
