### 2022-12-27 N-Wouda/ALNS#129 @leonlan
Packages (mostly Python)
- https://github.com/DEAP/deap
- https://github.com/jMetal/jMetalPy
- https://github.com/jmejia8/Metaheuristics.jl
- https://github.com/ac-tuwien/pymhlib
  - See https://github.com/ac-tuwien/pymhlib/blob/master/pymhlib/alns.py for alns
- https://github.com/thieu1995/mealpy
- https://github.com/ljvmiranda921/pyswarms
- https://github.com/anyoptimization/pymoo
- https://github.com/nojhan/paradiseo

Not a package but relevant
- https://github.com/alberto-santini/adaptive-large-neighbourhood-search
