### 2022-12-27 #5028 @skadio
Glad to hear that you find it useful, and thank you both again for the library!

Inspired by your alpha-UCB policy in ALNS, let me add one potential integration opportunity for future. This comment is not related to publication at JOSS, so feel free to discard.

===

Similar to the UCB policy, other Bandit policies can be considered when selecting operators, for instance Thompson Samling.

In that regard, the [MABWiser library](https://github.com/fidelity/mabwiser#quick-start) that offers a comprehensive set of policies would neatly work as the Selection operator. Disclaimer: the library is built within our group in the AI Center at Fidelity.

The idea is to enable the "select" parameter to be MABWiser object which implements a variety of Bandit policies.

So the current UCB policy

```
select = AlphaUCB(scores=[5, 2, 1, 0.5], alpha=0.05), num_repair, num_destroy)
```

becomes

```
select = AlphaUCB(scores=[5, 2, 1, 0.5], MABWiser(arms=num_repair * num_destroy, LearningPolicy.UCB1(alpha=0.05)
```

Under the hood, ALNS leverages the expected reward and arm prediction computations from MABWiser.

In principle, the user can select _**any available bandit policy**_. In that sense, AlphaUCB becomes `BanditALNS(scores, MABWiser)`

Btw, it might be worthwhile considering Thompson Sampling (TS) Bandits which is a common counterpart to UCB policies, and often outperforms it.

It would look like this:

```
select = BanditALNS(scores=[1, 1, 1, 0], MABWiser(arms=num_repair * num_destroy, LearningPolicy.ThompsonSampling())
```

The reward for TS is binary. Here I used 1 for best, better, and accepted solutions and 0 for rejected solutions. The 0/1 scoring schema for TS would be easier to tune than UCB where it is not clear how to pick _reasonable_ defaults.

Finally, both UCB and LS are **non-contextual** policies. Moving beyond them, one could also consider **contextual policies**, where a set of features that describes that current state (neighborhood) help decide the arm selection. I am not aware of any research that considers contextual bandit policies for ALNS --this would be super cool! MABWiser supports all of these, and more.
