### 2022-12-28 N-Wouda/ALNS#129 @leonlan
I think the list above includes the most relevant packages.

In this [branch](https://github.com/N-Wouda/ALNS/compare/joss-paper...state-of-the-field), I slightly modified the text in the "statement of need" and added a single line of text to refer to other libraries. This is what many JOSS papers seem to do as well, and IMO it feels enough.

It feels a bit of a stretch to write a section on "State of the Field", because none of these packages are completely dedicated to single trajectory and single objective algorithms. But I think I can also make an attempt to write this.

I'll have a better look at this tomorrow, I need to fix my jetlag :-)
