### 2022-12-28 28e6c33
Add new text and references
