+++ paper/paper.md
@@ -36,7 +36,10 @@ Because of this success, we expect the package to be useful to the wider operati
 It is common in the operations research community to re-implement heuristics [@Swan_et_al:2022; @vidal:2022].
 As a result, implementations of the ALNS algorithm are relatively limited, are typically tied tightly to one particular problem domain, and often implement just a single acceptance criterion and operator selection scheme.
 The survey of @Windras_Mara_et_al:2022 corroborates these claims: 205 out of the 251 papers they survey only consider a simulated annealing acceptance criterion, and only one paper uses an operator selection scheme that is not based on the roulette wheel mechanism of @Ropke_Pisinger:2006.
-While many general-purpose libraries exists that could facilitate the implementation of metaheuristics, such as ..., these libraries often do not include a complete implementation of the ALNS algorithm.
+A large number of software libraries exists that facilitate the implementation of metaheuristics [@Parejo2012-MetaheuristicOptimizationFrameworks]. 
+Several widely-used libraries are `DEAP` [@Fortin2012-DEAPEvolutionaryAlgorithms], `ECJ` [@Scott2019-ECJ20General], `jMetal` [@Durillo2011-JMetalJavaFramework], `Metaheuristics.jl` [@Mejia-de-Dios2022-MetaheuristicsJuliaPackage], and `Paradiseo` [@Dreo2021-ParadiseoModularFramework]. 
+These libraries include implementations of a wide range metaheuristics, such as evolutionary algorithms, particle swarm optimization, and iterated local search.
+However, none of the aforementioned libraries provide a complete implementation of the ALNS algorithm out-of-the-box. 
 This inhibits experimentation with different aspects of the algorithm, and makes re-use by others or in other problem domains difficult.
 Our `alns` package, by contrast, offers a clear and problem-agnostic API for using the ALNS algorithm, and provides many acceptance criteria and operator selection schemes.
 Additionally, we provide diagnostic statistics, plotting methods, logging, and the ability to register custom callbacks at various points of the search.
