+++ docs/source/index.rst
@@ -4,7 +4,6 @@ ALNS
 ``alns`` is a general, well-documented and tested implementation of the adaptive large neighbourhood search (ALNS) metaheuristic in Python.
 ALNS is an algorithm that can be used to solve difficult combinatorial optimisation problems.
 The algorithm begins with an initial solution.
-This solution should be feasible, but need not be very good.
 Then the algorithm iterates until a stopping criterion is met.
 In each iteration, a destroy and repair operator are selected, which transform the current solution into a candidate solution.
 This candidate solution is then evaluated by an acceptance criterion, and the operator selection scheme is updated based on the evaluation outcome.
