+++ paper/paper.md
@@ -37,9 +37,9 @@ It is common in the operations research community to re-implement heuristics [@S
 As a result, implementations of the ALNS algorithm are relatively limited, are typically tied tightly to one particular problem domain, and often implement just a single acceptance criterion and operator selection scheme.
 The survey of @Windras_Mara_et_al:2022 corroborates these claims: 205 out of the 251 papers they survey only consider a simulated annealing acceptance criterion, and only one paper uses an operator selection scheme that is not based on the roulette wheel mechanism of @Ropke_Pisinger:2006.
 A large number of software libraries exists that facilitate the implementation of metaheuristics [@Parejo2012-MetaheuristicOptimizationFrameworks]. 
-Several widely-used libraries are `DEAP` [@Fortin2012-DEAPEvolutionaryAlgorithms], `ECJ` [@Scott2019-ECJ20General], `jMetal` [@Durillo2011-JMetalJavaFramework], `Metaheuristics.jl` [@Mejia-de-Dios2022-MetaheuristicsJuliaPackage], and `Paradiseo` [@Dreo2021-ParadiseoModularFramework]. 
-These libraries include implementations of a wide range metaheuristics, such as evolutionary algorithms, particle swarm optimization, and iterated local search.
-However, none of the aforementioned libraries provide a complete implementation of the ALNS algorithm out-of-the-box. 
+Several widely-used packages are `DEAP` [@Fortin2012-DEAPEvolutionaryAlgorithms], `ECJ` [@Scott2019-ECJ20General], `jMetal` [@Durillo2011-JMetalJavaFramework], `Metaheuristics.jl` [@Mejia-de-Dios2022-MetaheuristicsJuliaPackage], and `Paradiseo` [@Dreo2021-ParadiseoModularFramework]. 
+However, these libraries generally focus on evolutionary algorithms and multi-objective optimization, and do not provide a complete implementation of ALNS algorithm out of the box. 
+Although it should be possible to adapt available algorithms to resemble the ALNS algorithm, it requires additional effort from users to add interfaces specific to ALNS, such as operator selection schemes and acceptance criteria. 
 This inhibits experimentation with different aspects of the algorithm, and makes re-use by others or in other problem domains difficult.
 Our `alns` package, by contrast, offers a clear and problem-agnostic API for using the ALNS algorithm, and provides many acceptance criteria and operator selection schemes.
 Additionally, we provide diagnostic statistics, plotting methods, logging, and the ability to register custom callbacks at various points of the search.
