+++ paper/paper.md
@@ -45,6 +45,18 @@ Our `alns` package, by contrast, offers a clear and problem-agnostic API for usi
 Additionally, we provide diagnostic statistics, plotting methods, logging, and the ability to register custom callbacks at various points of the search.
 These allow researchers and practitioners to rapidly develop state-of-the-art metaheuristics in a wide range of problem domains.
 
+
+# State of the field
+
+A large number of software libraries exists that facilitate the implementation of metaheuristics [@Parejo2012-MetaheuristicOptimizationFrameworks].
+The most widely-used packages generally focus on population-based evolutionary algorithms and multi-objective optimization, for example `DEAP` [@Fortin2012-DEAPEvolutionaryAlgorithms], `ECJ` [@Scott2019-ECJ20General], `jMetal` [@Durillo2011-JMetalJavaFramework] and `Metaheuristics.jl` [@Mejia-de-Dios2022-MetaheuristicsJuliaPackage]. 
+These libraries provide limited functionality for single-trajectory and single-objective algorithms, which is the class of algorithms that ALNS belongs to.
+Among libraries that focus on single-trajectory algorithms are `Paradiseo`, which is implemented in C++, and `Chips-n-Salsa` [@Cicirello2020-ChipsnSalsaJavaLibrary], which is implemented in Java.
+These libraries support algorithms such as simulated annealing, tabu search, and iterated local search, but do not include an implementation of ALNS.
+While it is possible to adapt the available algorithms to mimic the ALNS algorithm, it would require additional effort from users to add features specific to ALNS, such as operator selection schemes.
+Our `alns` package provides a feature-complete implementation of ALNS, is implemented in Python, and also supports other single-trajectory as special cases. 
+
+
 # Features
 
 At its core, ALNS is an iterative ruin-and-recreate algorithm that runs until some stopping criterion is met.
