+++ paper/paper.bib
@@ -211,3 +211,20 @@
   doi = {10.1007/s00500-011-0754-8},
   language = {en},
 }
+
+@article{Cicirello2020-ChipsnSalsaJavaLibrary,
+  title = {Chips-n-{{Salsa}}: {{A Java Library}} of {{Customizable}}, {{Hybridizable}}, {{Iterative}}, {{Parallel}}, {{Stochastic}}, and {{Self-Adaptive Local Search Algorithms}}},
+  shorttitle = {Chips-n-{{Salsa}}},
+  author = {Cicirello, Vincent A.},
+  year = {2020},
+  month = aug,
+  journal = {Journal of Open Source Software},
+  volume = {5},
+  number = {52},
+  pages = {2448},
+  issn = {2475-9066},
+  doi = {10.21105/joss.02448},
+  abstract = {Cicirello, V. A., (2020). Chips-n-Salsa: A Java Library of Customizable, Hybridizable, Iterative, Parallel, Stochastic, and Self-Adaptive Local Search Algorithms. Journal of Open Source Software, 5(52), 2448, https://doi.org/10.21105/joss.02448},
+  language = {en},
+  file = {/Users/leonlan/Zotero/storage/SBYSY9VD/Cicirello2020_Chips-n-Salsa.pdf}
+}
