+++ paper/paper.bib
@@ -228,3 +228,13 @@
   language = {en},
   file = {/Users/leonlan/Zotero/storage/SBYSY9VD/Cicirello2020_Chips-n-Salsa.pdf}
 }
+
+@misc{Santini2019-AdaptiveLargeNeighbourhood,
+  title = {Adaptive {{Large Neighbourhood Search}}},
+  author = {Santini, Alberto},
+  year = {2019},
+  publisher = {GitHub},
+  journal = {GitHub repository},
+  howpublished = {\url{hhttps://github.com/alberto-santini/adaptive-large-neighbourhood-search}},
+  abstract = {ALNS header-only library (loosely) based on the original implementation by Stefan Ropke.},
+}
