+++ paper/paper.bib
@@ -238,3 +238,19 @@
   howpublished = {\url{hhttps://github.com/alberto-santini/adaptive-large-neighbourhood-search}},
   abstract = {ALNS header-only library (loosely) based on the original implementation by Stefan Ropke.},
 }
+
+@article{Miranda2018-PySwarmsResearchToolkit,
+  title = {{{PySwarms}}: A Research Toolkit for {{Particle Swarm Optimization}} in {{Python}}},
+  shorttitle = {{{PySwarms}}},
+  author = {Miranda, Lester James},
+  year = {2018},
+  month = jan,
+  journal = {Journal of Open Source Software},
+  volume = {3},
+  number = {21},
+  pages = {433},
+  issn = {2475-9066},
+  doi = {10.21105/joss.00433},
+  abstract = {Miranda, (2018). PySwarms: a research toolkit for Particle Swarm Optimization in Python. Journal of Open Source Software, 3(21), 433, https://doi.org/10.21105/joss.00433},
+  language = {en},
+}
