### 2022-12-29 89622f9
Add a "getting started" code template (#131)
* Add a "getting started" code template

* Move data files into examples/data, update notebooks

* Add warning block about copying to ALNS.add_destroy_operator

* Reference copying, quickstart
