+++ README.md
@@ -3,58 +3,90 @@
 [![Documentation Status](https://readthedocs.org/projects/alns/badge/?version=latest)](https://alns.readthedocs.io/en/latest/?badge=latest)
 [![codecov](https://codecov.io/gh/N-Wouda/ALNS/branch/master/graph/badge.svg)](https://codecov.io/gh/N-Wouda/ALNS)
 
-This package offers a general, well-documented and tested
-implementation of the adaptive large neighbourhood search (ALNS)
-metaheuristic. It may be installed in the usual way as
+``alns`` is a general, well-documented and tested implementation of the adaptive
+large neighbourhood search (ALNS) metaheuristic in Python. ALNS is an algorithm
+that can be used to solve difficult combinatorial optimisation problems. The
+algorithm begins with an initial solution. This solution should be feasible, but
+need not be very good. Then the algorithm iterates until a stopping criterion is
+met. In each iteration, a destroy and repair operator are selected, which
+transform the current solution into a candidate solution. This candidate
+solution is then evaluated by an acceptance criterion, and the operator
+selection scheme is updated based on the evaluation outcome.
+
+It may be installed in the usual way as
+
 ```
 pip install alns
 ```
 
+The documentation is available [here][1].
+
+### Getting started
+
+The `alns` library provides the ALNS algorithm and various acceptance criteria,
+operator selection schemes, and stopping criteria. To solve your own problem,
+you should provide the following:
+
+- A solution state for your problem that implements an `objective()` function.
+- An initial solution.
+- One or more destroy and repair operators tailored to your problem.
+
+A "quickstart" code template is available [here][10].
+
 ### Examples
-If you wish to dive right in, the [documentation][1] contains example notebooks
-showing how the ALNS library may be used. These include:
+
+We provide several example notebooks showing how the ALNS library may be used.
+These include:
 
 - The travelling salesman problem (TSP), [here][2]. We solve an instance of 131
-  cities in one minute to a 2% optimality gap, using very simple destroy and
-  repair heuristics.
-- The capacitated vehicle routing problem (CVRP), [here][8]. We solve an 
-  instance with 241 customers to within 3% of optimality using a combination
-  of a greedy repair operator, and a _slack-induced substring removal_ destroy
-  operator.
-- The cutting-stock problem (CSP), [here][4]. We solve an instance with
-  180 beams over 165 distinct sizes to within 1.35% of optimality in
-  only a very limited number of iterations.
-- The resource-constrained project scheduling problem (RCPSP), [here][6]. We solve 
-  an instance with 90 jobs and 4 resources to within 4% of the best known solution,
-  using a number of different operators and enhancement techniques from the 
-  literature.
+  cities using very simple destroy and repair heuristics.
+- The capacitated vehicle routing problem (CVRP), [here][8]. We solve an
+  instance with 241 customers using a combination of a greedy repair operator,
+  and a _slack-induced substring removal_ destroy operator.
+- The cutting-stock problem (CSP), [here][4]. We solve an instance with 180
+  beams over 165 distinct sizes in only a very limited number of iterations.
+- The resource-constrained project scheduling problem (RCPSP), [here][6]. We
+  solve an instance with 90 jobs and 4 resources using a number of different
+  operators and enhancement techniques from the literature.
 - The permutation flow shop problem (PFSP), [here][9]. We solve an instance with
-  50 jobs and 20 machines to within 1.5% of the best known solution. Moreover,
-  we demonstrate multiple advanced features of ALNS, including auto-fitting the
-  acceptance criterion and adding local search to repair operators. We also
-  demonstrate how one could tune ALNS parameters.
[…]
