+++ paper/paper.bib
@@ -119,3 +119,138 @@
     year={2022},
     pages={393–406},
 }
+
+@article{Fortin2012-DEAPEvolutionaryAlgorithms,
+  title = {{{DEAP}}: {{Evolutionary Algorithms Made Easy}}},
+  shorttitle = {{{DEAP}}},
+  author = {Fortin, F{\'e}lix-Antoine and Rainville, Fran{\c c}ois-Michel De and Gardner, Marc-Andr{\'e} and Parizeau, Marc and Gagn{\'e}, Christian},
+  year = {2012},
+  journal = {Journal of Machine Learning Research},
+  volume = {13},
+  number = {70},
+  pages = {2171--2175},
+  issn = {1533-7928},
+  abstract = {DEAP is a novel evolutionary computation framework for rapid prototyping and testing of ideas. Its design departs from most other existing frameworks in that it seeks to make algorithms explicit and data structures transparent, as opposed to the more common black-box frameworks. Freely available with extensive documentation at http://deap.gel.ulaval.ca, DEAP is an open source project under an LGPL license.},
+}
+
+@inproceedings{Scott2019-ECJ20General,
+  title = {{{ECJ}} at 20: Toward a General Metaheuristics Toolkit},
+  shorttitle = {{{ECJ}} at 20},
+  booktitle = {Proceedings of the {{Genetic}} and {{Evolutionary Computation Conference Companion}}},
+  author = {Scott, Eric O. and Luke, Sean},
+  year = {2019},
+  month = jul,
+  series = {{{GECCO}} '19},
+  pages = {1391--1398},
+  publisher = {{Association for Computing Machinery}},
+  address = {{New York, NY, USA}},
+  doi = {10.1145/3319619.3326865},
+  abstract = {ECJ is now 20 years old. Begun as a genetic programming and evolutionary computation library in Java, it has since established itself as historically one of the most popular EC toolkits worldwide. In 2016 we received a National Science Foundation grant to improve ECJ in many ways with an eye toward making it a useful toolkit not just for EC but for the broader metaheuristics community. This paper is a report on our efforts to this end. We discuss new metaheuristics frameworks and representations added to ECJ and the design challenges that they raise for a general-purpose framework, as well as testing facilities and other support tools. We conclude with our future directions for the library.},
+  isbn = {978-1-4503-6748-6},
+}
+
+@article{Durillo2011-JMetalJavaFramework,
+  title = {{{jMetal}}: {{A Java}} Framework for Multi-Objective Optimization},
+  shorttitle = {{{jMetal}}},
+  author = {Durillo, Juan J. and Nebro, Antonio J.},
+  year = {2011},
+  month = oct,
+  journal = {Advances in Engineering Software},
+  volume = {42},
+  number = {10},
+  pages = {760--771},
+  issn = {0965-9978},
+  doi = {10.1016/j.advengsoft.2011.05.014},
[…]
