+++ paper/paper.bib
@@ -226,7 +226,6 @@
   doi = {10.21105/joss.02448},
   abstract = {Cicirello, V. A., (2020). Chips-n-Salsa: A Java Library of Customizable, Hybridizable, Iterative, Parallel, Stochastic, and Self-Adaptive Local Search Algorithms. Journal of Open Source Software, 5(52), 2448, https://doi.org/10.21105/joss.02448},
   language = {en},
-  file = {/Users/leonlan/Zotero/storage/SBYSY9VD/Cicirello2020_Chips-n-Salsa.pdf}
 }
 
 @misc{Santini2019-AdaptiveLargeNeighbourhood,
@@ -235,7 +234,7 @@
   year = {2019},
   publisher = {GitHub},
   journal = {GitHub repository},
-  howpublished = {\url{hhttps://github.com/alberto-santini/adaptive-large-neighbourhood-search}},
+  howpublished = {\url{https://github.com/alberto-santini/adaptive-large-neighbourhood-search}},
   abstract = {ALNS header-only library (loosely) based on the original implementation by Stefan Ropke.},
 }
 
