+++ paper/paper.bib
@@ -57,28 +57,16 @@
     year = {2022},
 }
 
-@article{Wouda_et_al:2022,
+@article{Wouda_et_al:2023,
     title={An adaptive large neighbourhood search metaheuristic for hourly learning activity planning in personalised learning},
     ISSN={0305-0548},
     DOI={10.1016/j.cor.2022.106089},
     abstractNote={Personalised learning offers an alternative method to one-size-fits-all education in schools, and has seen increasing adoption over the past several years. Personalised learning’s focus on learner-driven education requires novel scheduling methods. In this paper we introduce the hourly, learner-driven activity planning problem of personalised learning, and formulate scheduling methods to solve it. We present an integer linear programming model of the problem, but this model does not generate schedules sufficiently quickly for use in practice. To overcome this, we propose an adaptive large neighbourhood search metaheuristic to solve the problem instead. The metaheuristic’s performance is compared against optimal solutions in a large numerical study of 14,400 instances. These instances are representative of secondary education in the Netherlands, and were developed from expert opinions. Solutions on average deviate only 1.6% from optimal results. Further, our experiments numerically demonstrate the mitigating effects changes to the structure and staffing of secondary education have on the challenges of satisfying learner instruction demands in personalised learning.},
     journal={Computers & Operations Research},
     author={Wouda, Niels A. and Aslan, Ayse and Vis, Iris F. A.},
-    year={2022},
+    year={2023},
+    volume = {151},
     pages={106089},
-}
-
-@article{vidal:2022,
-    title = {Hybrid genetic search for the CVRP: Open-source implementation and SWAP* neighborhood},
-    journal = {Computers & Operations Research},
-    volume = {140},
-    pages = {105643},
-    year = {2022},
-    issn = {0305-0548},
-    doi = {10.1016/j.cor.2021.105643},
-    url = {https://www.sciencedirect.com/science/article/pii/S030505482100349X},
-    author = {Thibaut Vidal},
-    abstract = {The vehicle routing problem is one of the most studied combinatorial optimization topics, due to its practical importance and methodological interest. Yet, despite extensive methodological progress, many recent studies are hampered by the limited access to simple and efficient open-source solution methods. Given the sophistication of current algorithms, reimplementation is becoming a difficult and time-consuming exercise that requires extensive care for details to be truly successful. Against this background, we use the opportunity of this short paper to introduce a simple – open-source – implementation of the hybrid genetic search (HGS) specialized to the capacitated vehicle routing problem (CVRP). This state-of-the-art algorithm uses the same general methodology as Vidal et al. (2012) but also includes additional methodological improvements and lessons learned over the past decade of research. In particular, it includes an additional neighborhood called SWAP* which consists in exchanging two customers between different routes without an insertion in place. As highlighted in our study, an efficient exploration of SWAP* moves significantly contributes to the performance of local searches. Moreover, as observed in experimental comparisons with other recent approaches on the classical instances of Uchoa et al. (2017), HGS still stands as a leading metaheuristic regarding solution quality, convergence speed, and conceptual simplicity.}
 }
 
 @article{Dueck:1993,
@@ -232,7 +220,6 @@
   title = {Adaptive {{Large Neighbourhood Search}}},
   author = {Santini, Alberto},
   year = {2019},
-  publisher = {GitHub},
   journal = {GitHub repository},
   howpublished = {\url{https://github.com/alberto-santini/adaptive-large-neighbourhood-search}},
   abstract = {ALNS header-only library (loosely) based on the original implementation by Stefan Ropke.},
