### 2023-01-01 b6d8ade
Update template list for what a user needs to provide in docs
