### 2023-01-01 c131aa6
Updates from JOSS review
Specifically:
- Rephrase State/AC/SC/OSS docstrings to clarify subclassing a bit more.
- Remove/describe use of @dataclass in notebooks.
