+++ docs/source/conf.py
@@ -52,6 +52,7 @@ extensions = [
     "sphinx.ext.autodoc",
     # "sphinx.ext.autosummary",
     "sphinx.ext.intersphinx",
+    "sphinx_rtd_theme",
     "nbsphinx",
     "numpydoc",
 ]
