+++ docs/requirements.txt
@@ -4,3 +4,5 @@ tomli >= 2.0.1
 nbsphinx >= 0.8.9
 ipython >= 8.6.0
 numpydoc >= 1.5.0
+sphinx_rtd_theme >= 0.5.1
+docutils==0.16
