### 2023-01-01 N-Wouda/ALNS#135 @N-Wouda
Integration with optimization modeling/solver software
In https://github.com/openjournals/joss-reviews/issues/5028, @skadio wrote:

>How you considered the integration of ALNS library with Optimization Modeling/Solving technology? For example, ALNS is well-known to be great complement to Constraint Programming, see e.g., the famous P. Shaw paper from 1998 https://link.springer.com/chapter/10.1007/3-540-49481-2_30
>
>These ideas have found concrete implementations in libraries, e.g.,
>
> - In Choco Constraint Solver (in java) https://choco-solver.org/docs/solving/lns/#:~:text=LNS%20is%20a%20two%2Dphase,their%20value%20in%20the%20solution.
> - In Google OR-Tools Solver (available in Python) https://acrogenesis.com/or-tools/documentation/user_manual/manual/metaheuristics/jobshop_lns.html
> - And even as solver-independent LNS modeling languages e.g. MiniLNS (similar to the larger MiniZinc language/ecosystem)
https://people.eng.unimelb.edu.au/pstuckey/papers/MiniLNS.pdf
>
>Here, the ALNS part is "generic" whereas the problem specific details are left to the user. Beyond vanilla problems, where there are a lot of side constraints, the user might have trouble "modeling" the state, in the first place.
>
>What would be amazing is IF one could take ALNS, and then in the STATE description to use an off-the-shelf solver to model and solve the problem. This is probably easier said than done, but I believe there is a tremendous potential if we could build out the path for such integration.
>
>Let's take Google OR-Tools for instances, since it is a great performing solver, has a high-level modeling language with support for many interesting constraints and available in Python. I imagine the STATE can be the initial Constraint model. Notice that, by solving it (potentially without an objective function), IMMEDIATELY generates an initial solution. So this initial construction comes for free (alleviating the user further). In some problems, especially with many constraints, even finding this solution can be tricky. In the Documentation I noticed, some simple/clever tricks are used like setting the first variable to 1, for instance. Then the relax/repair operators need to work off of the solvers variables.
>
>I think this should be doable, and would give best of both worlds: ability to solve optimization problems at scale (thanks to local search) as opposed to exact solutions (which might not scale, e.g., mixed integer programs) BUT with the modeling support (as opposed to tradition local search where the user must implement everything from scratch).
>
>I think there is vital gap in the Optimization community for this, especially within the Python tech stack which also is the main technology powering Data Science/ML/AI community. Feel free to connect with me later on, if this research directions sounds interesting. I would be happy to brainstorm.
