### 2023-01-01 #5028 @N-Wouda
@skadio I'll respond in three separate parts, addressing first your minimum requested changes, then your detailed comments, and finally some of the more tangentially related items.

## Minimum changes

>The notebooks and documentations are great! Someone with an OR background should not have any problem using the library. That said, we need a bit more on the example usage for outsiders. With a quick low-hanging effort, the library can target a larger audience --which I believe is where largest potential is. To inspire the authors, I took a stab at this below. More on this in details below.

This could be very exciting! We have adopted your code template ([here](https://alns.readthedocs.io/en/latest/setup/template.html) in the documentation, along with some notes), and link to it from our paper and the repository readme.

>See the JOSS required item above to please add "State of the Field" to cover how this package compares to other packages. If none such exists (which I doubt it, since stochastic local search is well-known), please make the contribution explicit. If there are other tools, please cover mainstream ones, if any and mention where each would benefit the most (e.g., genetic algorithms for multi-point solutions vs. alns)

Thank you for pointing this out: we had initially missed this requirement. We now discuss several commonly used libraries in our “statement of need” section in our paper, and link to a recent review article for readers wishing to learn more.

>See the JOSS required item above to please add Community Guidelines, if not already.

These were a bit buried in the documentation. We have several pages in the documentation detailing 1) [how to contribute to the software](https://alns.readthedocs.io/en/latest/setup/contributing.html), and 2) [report issues/seek support](https://alns.readthedocs.io/en/latest/setup/getting_help.html). Links to these pages have been added to the repository readme, so they are easier to find for users.
We have also set-up issue and pull request templates in the repository, to guide people opening issues or pull requests.

## Detailed comments

> I believe, at a MINIMUM, the library requires the user to define:
> 1.	a STATE
> 2.	at least one DESTROY and one REPAIR operator, and
> 3.	an INITIAL STATE.

This is correct. We have made this more explicit in our paper (under “Features”), the repository readme, and in the documentation.

>The pydoc in `State.py` class reads "you are encouraged to subclass it". My understanding is, doing is MANDATORY not encourage. If so, please re-word. Same everywhere to be explicit about what is mandatory.

This is not required, but is something we used to do to satisfy the type checker. A better way to encode the actual constraint this is meant to satisfy---that a state object must have an objective() member function---is the use of the `Protocol` type. But that was [only added to Python in version 3.8](https://docs.python.org/3/library/typing.html#typing.Protocol), and the `alns` package supports Python 3.7 until its end-of-life ([in June 2023](https://endoflife.date/python)).

Luckily, none of this is problematic to the user. It will work with and without explicitly subclassing `alns.State`. To remove some of the confusion, we have removed this text from the `State.py` file, and also similarly rephrased the base classes for acceptance criteria, operator selection schemes, and stopping criteria.

>In the Destroy() operator is it mandatory to work with the deep copy object? It seems so to me, but if correct, please be explicit about it since this is critical, no? Again, this is where the boiler plate comes quite handy. I explicitly mentioned this in the code TODO above

The destroy operator is given the current solution maintained by the ALNS algorithm. It is indeed critical that this solution _itself_ is not modified, but instead a copy is changed/constructed. Since some solution states can be quite big, a deep copy might be very expensive – and there could be a problem-specific way to avoid having to deep copy a lot of data. As such, the `alns` package does not deep copy by default, but instead leaves this to the user.

We agree this was not phrased clearly. We have added a warning about this to the API documentation ([here](https://alns.readthedocs.io/en/latest/api/alns.html#alns.ALNS.ALNS.add_destroy_operator)), and link to it from the [quick start template](https://alns.readthedocs.io/en/latest/setup/template.html). We have also kept the relevant TODO in the quick start code.

>If I am following correctly, the library is built for minimization. Again, this is quite critical and needs to be explicitly stated. See my TODO comments about this and aim for something similar

This is correct. We have added a note about this to the [quick start template](https://alns.readthedocs.io/en/latest/setup/template.html) in the documentation, and explain there how one could also do maximisation.

>I saw some use of @DataClass annotation in some of the STATE definitions. Why is this needed/useful? and when? Documentation does not say much about its use. Please add some color

We mostly use the [`@dataclass`](https://docs.python.org/3/library/dataclasses.html#dataclasses.dataclass) decorator to avoid having to be explicit about the class’ data fields and the `__init__` method. This is not really something specific to `alns`, but rather something built into Python since 3.7. In the example notebooks, we have added a few notes here and there to explain how this can be helpful in simplifying the state implementation a bit.

>In the source code /examples folder, if you create a /data folder, then you can move all data files there. This would clean-up the structure and keep the notebooks (the most interesting bit) much more visible to library users.

This is a great idea! We have done exactly what you suggested, and this has really cleaned up the examples/ directory.

>On that note, here is nice to have comment, and completely not necessary, that you are also probably aware of: there is A LOT of great code used in the documentation. At the same time, there is an opportunity to have them built-in within the library. I am thinking of sklearn and other ML library that comes ready with data loaders, plotting classes, that can be imported and used directly. If these data elements were moved into the library, then you could drop all that code from the documentation and hugely simplify the ALNS presentation with nice imports (say draw_tsp()), and the purpose is clear from the context. This would also hide a lot non-ALNS related imports/configs from sight as well. Again, this is not critical at all. Just mentioning, and I am sure you already considered.

Thank you for your suggestion! One of our selling points is that our alns package is problem-agnostic, and could in principle be used to solve any problem having an objective and some constraints. As such, we have opted not to already include problem-specific data loaders or related tools. If, however, a community of users working on related problems (e.g., TSP or VRP) should emerge, we are very willing to support such uses better ‘out of the box’, from an `alns.tsp` or `alns.vrp` submodule. Such support could include data loaders, plotting facilities, and standard solution state classes, along with some destroy and repair operators.

## Other items

>How you considered the integration of ALNS library with Optimization Modeling/Solving technology?

We had not considered this before, but indeed, solving general mathematical programming problems should also be possible. There has been some work on this in recent years (e.g., [Hendel (2022)](https://link.springer.com/article/10.1007/s12532-021-00209-7)), but more can be done to simplify/improve the integration of heuristic methods for general optimization problems.

We think this is an interesting direction, and have copied your response to [a new issue in our repository](https://github.com/N-Wouda/ALNS/issues/132).

>Inspired by your alpha-UCB policy in ALNS, let me add one potential integration opportunity for future. This comment is not related to publication at JOSS, so feel free to discard.

Thank you for pointing us to the [MABWiser library](https://github.com/fidelity/mabwiser). Using MABs to select operators is a very exciting direction for ALNS, and integrating with your group’s work at Fidelity seems like a great way to very quickly get a large range of MABs going. We have copied your comment to [a new issue in our repository](https://github.com/N-Wouda/ALNS/issues/132), and will look into this more in the (near) future!
