### 2023-01-01 N-Wouda/ALNS#135 @N-Wouda
Solving general mathematical programming problems should also be possible. There has been some work on this in recent years (e.g., [Hendel (2022)](https://link.springer.com/article/10.1007/s12532-021-00209-7)), but more can be done to simplify/improve the integration of heuristic methods for general optimization problems.
