### 2023-01-06 #5028 @kenohori
Hello! First of all, congrats @N-Wouda on the nice library. Even as someone outside OR (but who implements a lot of optimisation problems), it was relatively easy to get it to create sample solutions to some simple problems. The paper, the code repository and the documentation are all easy to follow.

Mainly, I would like to second @skadio's comments on improving the help for non-OR users to bring their own problems into ALNS. This could make the library much more accessible to a wider audience. The code quickstart suggested by @skadio and the [ALNS features](https://alns.readthedocs.io/en/latest/examples/alns_features.html) page are good places to start, but I would have really appreciated an introduction to ALNS and some more general help (eg what problems it is good for, how to apply ALNS to your own problem, what is necessary to implement, how important different parts are).

Minor comments:

* A list of dependencies would be nice, even if it can be handled automatically by pip. From what I gathered, they are only matplotlib and numpy?
* Maybe I missed it, but is there a way to test that the library is working correctly? Unit tests would be best, but a script that uses all the main components of the library would be good as well.
