### 2023-01-06 236e307
Updates from JOSS review
