### 2023-01-06 #5028 @N-Wouda
> Hello! First of all, congrats @N-Wouda on the nice library. Even as someone outside OR (but who implements a lot of optimisation problems), it was relatively easy to get it to create sample solutions to some simple problems. The paper, the code repository and the documentation are all easy to follow.

Hello @kenohori! 👋 We are glad you managed to get started so quickly, and are grateful for your positive comments!

> Mainly, I would like to second @skadio's comments on improving the help for non-OR users to bring their own problems into ALNS. This could make the library much more accessible to a wider audience. The code quickstart suggested by @skadio and the [ALNS features](https://alns.readthedocs.io/en/latest/examples/alns_features.html) page are good places to start, but I would have really appreciated an introduction to ALNS and some more general help (eg what problems it is good for, how to apply ALNS to your own problem, what is necessary to implement, how important different parts are).

Thank you for this comment. I will write another page in the documentation explicitly introducing ALNS, explaining what type of problems it can be used to solve, and briefly describe what a minimal implementation should contain (in terms of ideas). This complements the quick start template, which is really about "so how do I turn those ideas into code". I expect I will have this finished later today, and will report back then.

> * A list of dependencies would be nice, even if it can be handled automatically by pip. From what I gathered, they are only matplotlib and numpy?

Yes, it's just `matplotlib` and `numpy`. We have added a short description of these dependencies to the repository readme and the documentation.

> * Maybe I missed it, but is there a way to test that the library is working correctly? Unit tests would be best, but a script that uses all the main components of the library would be good as well.

We have both! The tests are embedded in the `alns/` submodules: each submodule has a `tests/` directory that contains the (unit) tests. These are executed using `poetry run pytest` in the repository root: the [contributing](https://alns.readthedocs.io/en/latest/setup/contributing.html) page in the documentation explains how to set up poetry to get this running locally. The tests are also run every time we commit to a pull request or to the master branch, as the "Run pytest" step in the GitHub Actions workflow.

The test coverage is reported here: [codecov](https://app.codecov.io/gh/N-Wouda/ALNS).

Finally, the examples in the `examples/` directory are executable notebooks that each use all the main `alns` components. These can be run locally as well, and how to do that is explained [here](https://alns.readthedocs.io/en/latest/setup/installation.html#running-the-examples-locally).
