+++ docs/source/setup/introduction_to_alns.rst
@@ -1,0 +1,4 @@
+A brief introduction to ALNS
+============================
+
+TODO
