+++ docs/source/index.rst
@@ -24,9 +24,9 @@ It can be installed through *pip* via
 
 .. note::
 
-    Have a look at the :doc:`setup/template` to get started on your own ALNS metaheuristic.
-    If you are new to metaheuristics or ALNS, you might also benefit from reading :doc:`setup/introduction_to_alns`.
-    To set-up an installation from source, or to run the examples listed below yourself, please have a look at the :doc:`setup/installation`.
+    Have a look at the :doc:` quickstart template <setup/template>` to get started on your own ALNS metaheuristic.
+    If you are new to metaheuristics or ALNS, you might also benefit from reading the :doc:`introduction to ALNS <setup/introduction_to_alns>`.
+    To set-up an installation from source, or to run the examples listed below yourself, please have a look at the :doc:`installation instructions <setup/installation>`.
 
 .. toctree::
    :maxdepth: 1
