+++ docs/source/index.rst
@@ -22,11 +22,11 @@ It can be installed through *pip* via
 
    pip install alns
 
-.. note::
+.. hint::
 
-    Have a look at the :doc:` quickstart template <setup/template>` to get started on your own ALNS metaheuristic.
+    Have a look at the :doc:`quickstart template <setup/template>` to get started on your own ALNS metaheuristic.
     If you are new to metaheuristics or ALNS, you might also benefit from reading the :doc:`introduction to ALNS <setup/introduction_to_alns>`.
-    To set-up an installation from source, or to run the examples listed below yourself, please have a look at the :doc:`installation instructions <setup/installation>`.
+    To set up an installation from source, or to run the examples listed below yourself, please have a look at the :doc:`installation instructions <setup/installation>`.
 
 .. toctree::
    :maxdepth: 1
