### 2023-01-07 981645f
Cite handbook chapter, simplify a bit, and move operator stuff to template
