+++ docs/source/setup/introduction_to_alns.rst
@@ -5,7 +5,7 @@ Before we explain what ALNS is, we first need to describe the type of problems i
 These problems typically come from the field of `combinatorial optimisation <https://en.wikipedia.org/wiki/Combinatorial_optimization>`_, and have
 
 * A large (but finite) set of solutions (the *search space*)
-* An objective to be optimised by finding a *good* solution in the search space
+* An objective function to be optimised by finding a *good* solution in the search space
 
 A very well-known example of such a problem is the `travelling salesman problem <https://en.wikipedia.org/wiki/Travelling_salesman_problem>`_ (TSP), which asks
 
@@ -14,15 +14,14 @@ A very well-known example of such a problem is the `travelling salesman problem 
 It is easy to find *a* solution: :math:`1 \rightarrow 2 \rightarrow 3 \rightarrow \ldots \rightarrow n \rightarrow 1` is one option.
 Finding an *optimal* solution, however, can be very difficult since there are :math:`n!` possible solutions, and the search space thus grows very quickly with :math:`n`.
 
-Because of the explosive growth of the search space, combinatorial problems cannot be solved optimally by testing every solution in the search space.
-In such cases, a heuristic method can help find good (but not necessarily optimal) solutions quickly.
-ALNS is one such heuristic method.
-
 .. note::
 
-    :ref:`/examples/travelling_salesman_problem.ipynb` solves a TSP instance using ALNS.
+    :ref:`This example notebooks </examples/travelling_salesman_problem.ipynb>` solves a TSP instance using ALNS.
 
-Specifically, the adaptive large neighbourhood search metaheuristic is:
+Due to the explosive growth of the search space, most combinatorial problems cannot be solved optimally by testing every solution in the search space.
+For these problems, a heuristic method can help find good (but not necessarily optimal) solutions quickly.
+The adaptive large neighbourhood search (ALNS) algorithm is one such heuristic method.
+Specifically, ALNS is:
 
 * A large neighbourhood search (LNS) method.
   LNS methods explore large subsets of the search space in a systematic manner.
@@ -34,14 +33,19 @@ Specifically, the adaptive large neighbourhood search metaheuristic is:
 * Adaptive.
   During the search, ALNS learns about operators that are more effective than others, and uses that information to choose those effective operators more often.
 
+.. note::
+
+    For a more thorough introduction to LNS and LNS-based metaheuristics like ALNS, the handbook chapter of `Pisinger and Røpke (2019) <https://doi.org/10.1007/978-3-319-91086-4_4>`_ may be useful.
+
 ALNS begins with an initial solution and then iterates until a stopping criterion is met.
 In each iteration, a destroy and repair operator are selected, which transform the current solution into a candidate solution.
 This candidate solution is then evaluated by an acceptance criterion, and the operator selection scheme is updated based on the evaluation outcome.
+Once the stopping criterion is met, ALNS returns the best solution it has found.
 In pseudocode, ALNS works as follows:
 
     .. line-block::
 
-        **Input:** an initial solution :math:`s`
+        **Input:** initial solution :math:`s`
         **Output:** optimised solution :math:`s^*`
         :math:`s^* \gets s`
         **repeat** until stopping criteria is met:
@@ -60,9 +64,6 @@ You need to provide:
 * An initial solution.
 * One or more destroy and repair operators tailored to your problem.
 
-Typically, a good start for a destroy operator is *random removal*, which randomly destroys some part of the current solution.
-A good repair operator is *greedy repair*, which repairs the partially destroyed solution in a greedy manner.
-
 .. hint::
 
-    Now that you know a little more about the ideas behind ALNS, the :doc:`quickstart code template <template>` is a great place to get started using the ``alns`` package!
+    The :doc:`quickstart code template <template>` explains how to implement these in the format ``alns`` expects.
