+++ docs/source/setup/template.rst
@@ -13,6 +13,9 @@ You should provide the following:
   operator should copy the passed-in state; see
   :meth:`~alns.ALNS.ALNS.add_destroy_operator` for details.
 
+Typically, a good first destroy operator is *random removal*, which randomly destroys some part of the current solution.
+A good first repair operator is *greedy repair*, which repairs the partially destroyed solution in a greedy manner.
+
 .. note::
 
    The ``alns`` package assumes your problem is a minimisation problem. If you
@@ -83,7 +86,7 @@ The following is a quickstart template that can help you get started:
     best = result.best_state
     print(f"Best heuristic solution objective is {best.objective()}.")
 
-.. note::
+.. hint::
 
     Have a look at the examples to get a feeling for how to implement the TODOs
     in the quickstart template!
