### 2023-01-07 #5028 @N-Wouda
>I expect I will have this finished later today, and will report back then.

@kenohori I had less time yesterday than I had hoped, but the new _Introduction to ALNS_ page is finished, and available [here](https://alns.readthedocs.io/en/latest/setup/introduction_to_alns.html) in the documentation. The page explains the kind of problems ALNS can solve, and how ALNS works. We also link to a handbook chapter that provides further details, should a user be interested in exploring these methods further.

We have also updated the quick start page [here](https://alns.readthedocs.io/en/latest/setup/template.html) a little to provide suggestions for a first destroy and repair operator pair.
