### 2023-01-11 #5028 @kenohori
Great! Thanks, @N-Wouda. The page has all the basics and reads well.

After some numpy issues with Python 3.10, I managed to get the test suite working in Python 3.9 and everything looks good.

As far as I'm concerned, it's ready to be accepted 🎉 .
