### 2023-01-19 #5028 @N-Wouda
@hugoledoux thanks for you kind words! We are happy to hear this is ready for publication. In response to your checklist:

- [x] I have created a tagged release with version `v5.0.3` of the package.
- [x] I have uploaded said tagged release to Zenodo, with DOI 10.5281/zenodo.7551604 (https://doi.org/10.5281/zenodo.7551604)
- [x] I have checked that the archival metadata (author list, title, and ORCIDs) match those in the JOSS paper.
