### 2023-01-19 09de45c
minor bib case changes
