+++ paper/paper.bib
@@ -109,7 +109,7 @@
 }
 
 @article{Fortin2012-DEAPEvolutionaryAlgorithms,
-  title = {{{DEAP}}: {{Evolutionary Algorithms Made Easy}}},
+  title = {{DEAP}: Evolutionary Algorithms Made Easy},
   shorttitle = {{{DEAP}}},
   author = {Fortin, F{\'e}lix-Antoine and Rainville, Fran{\c c}ois-Michel De and Gardner, Marc-Andr{\'e} and Parizeau, Marc and Gagn{\'e}, Christian},
   year = {2012},
@@ -138,7 +138,7 @@
 }
 
 @article{Durillo2011-JMetalJavaFramework,
-  title = {{{jMetal}}: {{A Java}} Framework for Multi-Objective Optimization},
+  title = {{jMetal: A {J}ava Framework for Multi-Objective Optimization},
   shorttitle = {{{jMetal}}},
   author = {Durillo, Juan J. and Nebro, Antonio J.},
   year = {2011},
@@ -154,9 +154,9 @@
 }
 
 @inproceedings{Dreo2021-ParadiseoModularFramework,
-  title = {Paradiseo: From a Modular Framework for Evolutionary Computation to the Automated Design of Metaheuristics: 22 Years of {{Paradiseo}}},
+  title = {Paradiseo: From a Modular Framework for Evolutionary Computation to the Automated Design of Metaheuristics: 22 Years of {P}aradiseo},
   shorttitle = {Paradiseo},
-  booktitle = {Proceedings of the {{Genetic}} and {{Evolutionary Computation Conference Companion}}},
+  booktitle = {Proceedings of the Genetic and Evolutionary Computation Conference Companion},
   author = {Dreo, Johann and Liefooghe, Arnaud and Verel, S{\'e}bastien and Schoenauer, Marc and Merelo, Juan J. and Quemy, Alexandre and Bouvier, Benjamin and Gmys, Jan},
   year = {2021},
   month = jul,
@@ -170,7 +170,7 @@
 }
 
 @article{Mejia-de-Dios2022-MetaheuristicsJuliaPackage,
-  title = {Metaheuristics: {{A Julia Package}} for {{Single-}} and {{Multi-Objective Optimization}}},
+  title = {Metaheuristics: A {J}ulia Package for Single- and Multi-Objective Optimization},
   shorttitle = {Metaheuristics},
   author = {{Mej{\'i}a-de-Dios}, Jes{\'u}s-Adolfo and {Mezura-Montes}, Efr{\'e}n},
   year = {2022},
@@ -201,7 +201,7 @@
 }
 
 @article{Cicirello2020-ChipsnSalsaJavaLibrary,
-  title = {Chips-n-{{Salsa}}: {{A Java Library}} of {{Customizable}}, {{Hybridizable}}, {{Iterative}}, {{Parallel}}, {{Stochastic}}, and {{Self-Adaptive Local Search Algorithms}}},
+  title = {Chips-n-{S}alsa: A {J}ava Library of Customizable, Hybridizable, Iterative, Parallel, Stochastic, and Self-Adaptive Local Search Algorithms},
   shorttitle = {Chips-n-{{Salsa}}},
   author = {Cicirello, Vincent A.},
   year = {2020},
@@ -217,7 +217,7 @@
 }
 
 @misc{Santini2019-AdaptiveLargeNeighbourhood,
-  title = {Adaptive {{Large Neighbourhood Search}}},
+  title = {Adaptive Large Neighbourhood Search},
   author = {Santini, Alberto},
   year = {2019},
   journal = {GitHub repository},
@@ -226,7 +226,7 @@
 }
 
[…]
