### 2023-01-20 #5028 @N-Wouda
@skadio and @kenohori: thank you for your thorough reviews! I think the paper and particularly the documentation have been much improved as a result.

@hugoledoux and @danielskatz: thank you for your editorial work. It was a pleasure to publish in JOSS!
